requirements.txt
gradio
ultralytics
opencv-python
pillow
torch
torchvision